first official minor update. moved types and files into each separate folders
