+++ .changeset/hungry-ants-wash.md
@@ -1,0 +1,5 @@
+---
+"paystackly": minor
+---
+
+first official minor update. moved types and files into each separate folders
