+++ package.json
@@ -11,12 +11,13 @@
   ],
   "private": false,
   "dependencies": {
+    "axios": "^1.4.0",
     "dotenv": "^16.3.1"
   },
   "scripts": {
     "test": "nyc mocha --timeout 20000 --recursive dist/__tests__/**  --exit",
     "dev": "npm run clean & tsc -v && tsc -p tsconfig.json -w",
-    "build": "tsc -p tsconfig.json",
+    "build": "npm run clean & tsc -v & tsc -p tsconfig.json",
     "lint": "tsc",
     "ci": "npm run lint && npm run test && npm run build",
     "release": "npm run ci && changeset version && changeset publish",
